requirements.txt
django==1.9.8
pillow
django-pagedown
django-markdown-deux
django-crispy-forms
djangorestframework
